move sql logic to sqlite helper
