+++ crates/opsml_sql/src/base.rs
@@ -6,6 +6,7 @@ use crate::schemas::schema::{
 use async_trait::async_trait;
 use opsml_error::error::SqlError;
 use opsml_settings::config::OpsmlDatabaseSettings;
+use opsml_utils::semver::VersionParser;
 use std::fmt;
 pub enum CardSQLTableNames {
     Data,
@@ -34,6 +35,52 @@ impl fmt::Display for CardSQLTableNames {
         };
         write!(f, "{}", table_name)
     }
+}
+
+pub fn add_version_bounds(builder: &mut String, version: &str) -> Result<(), SqlError> {
+    let version_bounds = VersionParser::get_version_to_search(version)
+        .map_err(|e| SqlError::VersionError(format!("{}", e)))?;
+
+    // construct lower bound (already validated)
+    builder.push_str(
+        format!(
+            " AND (major >= {} AND minor >= {} and patch >= {})",
+            version_bounds.lower_bound.major,
+            version_bounds.lower_bound.minor,
+            version_bounds.lower_bound.patch
+        )
+        .as_str(),
+    );
+
+    if !version_bounds.no_upper_bound {
+        // construct upper bound based on number of components
+        if version_bounds.num_parts == 1 {
+            builder
+                .push_str(format!(" AND (major < {})", version_bounds.upper_bound.major).as_str());
+        } else if version_bounds.num_parts == 2
+            || version_bounds.num_parts == 3 && version_bounds.parser_type == VersionParser::Tilde
+            || version_bounds.num_parts == 3 && version_bounds.parser_type == VersionParser::Caret
+        {
+            builder.push_str(
+                format!(
+                    " AND (major == {} AND minor < {})",
+                    version_bounds.upper_bound.major, version_bounds.upper_bound.minor
+                )
+                .as_str(),
+            );
+        } else {
+            builder.push_str(
+                format!(
+                    " AND (major == {} AND minor == {} AND patch < {})",
+                    version_bounds.upper_bound.major,
+                    version_bounds.upper_bound.minor,
+                    version_bounds.upper_bound.patch
+                )
+                .as_str(),
+            );
+        }
+    }
+    Ok(())
 }
 
 #[async_trait]
